# apps/wca-data/openapi/wca-data-v1.yaml
openapi: 3.0.3
info:
  title: Unofficial WCA Data API
  version: 0.1.0
  description: |
    # Introduction

    Welcome to the **unofficial WCA Data API** for speedcube.com.br.

    This API provides public JSON access to World Cube Association competition data: events, countries, competitions, competitors, rankings, results, and scrambles. It is designed for applications that need a stable HTTP API backed by the public WCA Results Export without executing downloaded SQL or coupling data ingestion to user requests.

    > This API is not affiliated with, endorsed by, or operated by the World Cube Association.

    ## Data source

    Data is imported from the public **World Cube Association Results Export**. The import worker downloads the official TSV export, loads it into staging tables, transforms it into canonical PostgreSQL tables, and atomically publishes a dataset version when the import succeeds.

    The public API reads only from the active dataset in our database. It does not call the WCA during API requests.

    ## Freshness and availability

    The WCA export is not realtime. Results, rankings, and competition metadata reflect the last successfully published dataset. Use `GET /wca-data/v1/status` to inspect the active dataset, export date, scheduler configuration, and the most recent import run.

    If no active dataset exists, data endpoints return `503 dataset_unavailable`. This is intentional: production never falls back to fixture data.

...

...

    ```json
...

...

...

...

...
servers:
...
paths:
...
  /wca-data/v1/events:
    get:
      tags: [Reference data]
      summary: List WCA events
      operationId: listEvents
      description: |
        Lists WCA event identifiers and display names. Use `id` values such as `333`, `222`, or `clock` when filtering rankings, competitions, or results.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: WCA events.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventListResponse'
              examples:
                events:
                  value:
                    data:
                      - id: '333'
                        name: 3x3x3 Cube
                        format: time
                      - id: clock
                        name: Clock
                        format: time
                    pagination:
                      page: 1
                      pageSize: 50
                      total: 2
                      hasNextPage: false
                    meta:
                      datasetId: 2026-06-30-v2.0.2
                      exportDate: '2026-06-30T00:00:16.000Z'
                      exportVersion: v2.0.2
                      source: World Cube Association Results Export
        '503':
          $ref: '#/components/responses/DatasetUnavailable'
      x-codeSamples:
        - lang: cURL
          source: |
            curl 'http://speedcube.com.br/api/wca-data/v1/events?pageSize=25'
        - lang: JavaScript
          source: |
            const response = await fetch('http://speedcube.com.br/api/wca-data/v1/events?pageSize=25');
            const events = await response.json();
  /wca-data/v1/countries:
    get:
      tags: [Reference data]
      summary: List WCA countries
      operationId: listCountries
      description: |
        Lists WCA country records and ISO2 codes. Use `iso2Code` values when filtering competitions, persons, or country rankings.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: WCA countries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryListResponse'
              examples:
                countries:
                  value:
                    data:
                      - iso2Code: PL
                        name: Poland
                        continentId: europe
                      - iso2Code: US
                        name: United States
                        continentId: north-america
                    pagination:
                      page: 1
                      pageSize: 50
                      total: 2
                      hasNextPage: false
                    meta:
                      datasetId: 2026-06-30-v2.0.2
                      exportDate: '2026-06-30T00:00:16.000Z'
                      exportVersion: v2.0.2
                      source: World Cube Association Results Export
        '503':
          $ref: '#/components/responses/DatasetUnavailable'
  /wca-data/v1/continents:
    get:
      tags: [Reference data]
      summary: List WCA continents
      operationId: listContinents
      description: |
        Lists WCA continent identifiers used by country records and regional ranking filters. The identifiers are the raw WCA export values, such as `_Europe` and `_South America`.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: WCA continents.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContinentListResponse'
              examples:
                continents:
                  value:
                    data:
                      - id: _Europe
                        name: Europe
                      - id: _South America
                        name: South America
                    pagination:
                      page: 1
                      pageSize: 50
                      total: 7
                      hasNextPage: false
                    meta:
                      datasetId: 2026-06-30-v2.0.2
                      exportDate: '2026-06-30T00:00:16.000Z'
                      exportVersion: v2.0.2
                      source: World Cube Association Results Export
        '503':
          $ref: '#/components/responses/DatasetUnavailable'
  /wca-data/v1/formats:
    get:
      tags: [Reference data]
      summary: List WCA formats
      operationId: listFormats
      description: |
        Lists WCA round formats. Format records describe how many solves are expected, which solves are trimmed, and whether ranking is based on best, average, or another secondary value.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: WCA formats.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormatListResponse'
              examples:
                formats:
                  value:
                    data:
                      - id: a
                        name: Average of 5
                        shortName: Ao5
                        sortBy: average
                        sortBySecond: best
                        expectedSolveCount: 5
                        trimFastestN: 1
                        trimSlowestN: 1
                    pagination:
                      page: 1
                      pageSize: 50
                      total: 7
                      hasNextPage: false
                    meta:
                      datasetId: 2026-06-30-v2.0.2
                      exportDate: '2026-06-30T00:00:16.000Z'
                      exportVersion: v2.0.2
                      source: World Cube Association Results Export
        '503':
          $ref: '#/components/responses/DatasetUnavailable'
  /wca-data/v1/round-types:
    get:
      tags: [Reference data]
      summary: List WCA round types
      operationId: listRoundTypes
      description: |
        Lists WCA round type identifiers and display names. Result entries include the resolved round name, while this endpoint exposes the full reference table.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: WCA round types.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoundTypeListResponse'
              examples:
                roundTypes:
                  value:
                    data:
                      - id: f
                        name: Final
                        cellName: Final
                        isFinal: true
                    pagination:
                      page: 1
                      pageSize: 50
                      total: 11
                      hasNextPage: false
                    meta:
                      datasetId: 2026-06-30-v2.0.2
                      exportDate: '2026-06-30T00:00:16.000Z'
                      exportVersion: v2.0.2
                      source: World Cube Association Results Export
        '503':
          $ref: '#/components/responses/DatasetUnavailable'
  /wca-data/v1/championships:
    get:
      tags: [Reference data]
      summary: List WCA championships
      operationId: listChampionships
      description: |
        Lists championship annotations from the WCA export. Use `championshipType` to find world, continental, national, or special championship designations, and `competitionId` to inspect championship annotations for one competition.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: championshipType
          in: query
          description: Filter by championship type, for example `world`.
          schema:
            type: string
            example: world
        - name: competitionId
          in: query
          description: Filter by competition id.
          schema:
            type: string
            example: WorldChampionship2023
      responses:
        '200':
          description: WCA championship annotations.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChampionshipListResponse'
              examples:
                championships:
                  value:
                    data:
                      - id: 1
                        competitionId: WorldChampionship2023
                        championshipType: world
                    pagination:
                      page: 1
                      pageSize: 50
                      total: 1
                      hasNextPage: false
                    meta:
                      datasetId: 2026-06-30-v2.0.2
                      exportDate: '2026-06-30T00:00:16.000Z'
                      exportVersion: v2.0.2
                      source: World Cube Association Results Export
        '503':
          $ref: '#/components/responses/DatasetUnavailable'
  /wca-data/v1/competitions:
    get:
      tags: [Competitions]
      summary: List competitions
      operationId: listCompetitions
      description: |
        Lists competitions from the active dataset. Filters can be combined to narrow results by country, event, and year.

        Common use cases include listing all competitions in a country, finding competitions that held a given event, or browsing competitions for a calendar year.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: countryIso2
          in: query
          description: Filter competitions by WCA country ISO2 code.
          schema:
            type: string
            example: PL
        - name: eventId
          in: query
          description: Filter competitions that include a specific WCA event id.
          schema:
            type: string
            example: '333'
        - name: year
          in: query
          description: Filter competitions that start in this year.
          schema:
            type: integer
            minimum: 1982
            example: 2026
      responses:
        '200':
          description: Competition list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompetitionListResponse'
              examples:
                competitions:
                  value:
                    data:
                      - id: FixtureOpen2026
                        name: Fixture Open 2026
                        city: Warsaw
                        countryIso2: PL
                        information: Example competition loaded from WCA export data.
                        date:
                          start: '2026-06-30'
                          end: '2026-06-30'
                          numberOfDays: 1
                        events: ['333', clock]
                        delegates:
                          - name: Delegate Example
                            email: null
                        organizers:
                          - name: Organizer Example
                            email: null
                        venue:
                          name: Main venue
                          address: Warsaw, Poland
                          details: Main hall
                          coordinates:
                            latitude: 52.123456
                            longitude: 21.012345
                        externalWebsite: null
                        cancelled: false
                    pagination:
                      page: 1
                      pageSize: 50
                      total: 1
                      hasNextPage: false
                    meta:
                      datasetId: 2026-06-30-v2.0.2
                      exportDate: '2026-06-30T00:00:16.000Z'
                      exportVersion: v2.0.2
                      source: World Cube Association Results Export
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '503':
          $ref: '#/components/responses/DatasetUnavailable'
      x-codeSamples:
        - lang: cURL
          source: |
            curl 'http://speedcube.com.br/api/wca-data/v1/competitions?countryIso2=PL&eventId=333&year=2026'
        - lang: JavaScript
          source: |
            const url = new URL('http://speedcube.com.br/api/wca-data/v1/competitions');
            url.searchParams.set('countryIso2', 'PL');
            url.searchParams.set('eventId', '333');
            const response = await fetch(url);
            const competitions = await response.json();
  /wca-data/v1/competitions/{id}:
    get:
      tags: [Competitions]
      summary: Get competition by id
      operationId: getCompetition
      description: Returns a single competition by its WCA competition id.
      parameters:
        - name: id
          in: path
          required: true
          description: WCA competition id.
          schema:
            type: string
            example: WorldChampionship2023
      responses:
        '200':
          description: Competition record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompetitionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/DatasetUnavailable'
  /wca-data/v1/persons:
    get:
      tags: [Persons]
      summary: List persons
      operationId: listPersons
      description: |
        Lists WCA persons. Use `search` for name or id matching and `countryIso2` to restrict results to one WCA country.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: countryIso2
          in: query
          description: Filter persons by WCA country ISO2 code.
          schema:
            type: string
            example: US
        - name: search
          in: query
          description: Case-insensitive search by person name or WCA id.
          schema:
            type: string
            example: Park
      responses:
        '200':
          description: Person list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonListResponse'
              examples:
                persons:
                  value:
                    data:
                      - id: 2012PARK03
                        name: Max Park
                        countryIso2: US
                        gender: m
                    pagination:
                      page: 1
                      pageSize: 50
                      total: 1
                      hasNextPage: false
                    meta:
                      datasetId: 2026-06-30-v2.0.2
                      exportDate: '2026-06-30T00:00:16.000Z'
                      exportVersion: v2.0.2
                      source: World Cube Association Results Export
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '503':
          $ref: '#/components/responses/DatasetUnavailable'
      x-codeSamples:
        - lang: cURL
          source: |
            curl 'http://speedcube.com.br/api/wca-data/v1/persons?search=park'
  /wca-data/v1/persons/{id}:
    get:
      tags: [Persons]
      summary: Get person by WCA id
      operationId: getPerson
      description: Returns one public WCA person record by WCA id.
      parameters:
        - name: id
          in: path
          required: true
          description: WCA person id.
          schema:
            type: string
            example: 2012PARK03
      responses:
        '200':
          description: Person record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/DatasetUnavailable'
  /wca-data/v1/rankings:
    get:
      tags: [Rankings]
      summary: List rankings
      operationId: listRankings
      description: |
        Lists rankings for one event and ranking type. Set `region=world` for world rankings, `region=country` with `countryIso2` for country rankings, or `region=continent` with `continentId` for continent rankings.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: eventId
          in: query
          required: true
          description: WCA event id.
          schema:
            type: string
            example: '333'
        - name: type
          in: query
          required: true
          description: Ranking type.
          schema:
            type: string
            enum: [average, single]
            example: single
        - name: region
          in: query
          description: Ranking region. Defaults to `world` when omitted.
          schema:
            type: string
            enum: [continent, country, world]
            example: world
        - name: countryIso2
          in: query
          description: Required when `region=country`.
          schema:
            type: string
            example: US
        - name: continentId
          in: query
          description: Required when `region=continent`.
          schema:
            type: string
            example: europe
      responses:
        '200':
          description: Ranking entries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RankListResponse'
              examples:
                rankings:
                  value:
                    data:
                      - personId: 2012PARK03
                        eventId: '333'
                        type: single
                        region: world
                        best:
                          raw: 313
                        rank:
                          world: 1
                          continent: 1
                          country: 1
                          selected: 1
                    pagination:
                      page: 1
                      pageSize: 50
                      total: 1
                      hasNextPage: false
                    meta:
                      datasetId: 2026-06-30-v2.0.2
                      exportDate: '2026-06-30T00:00:16.000Z'
                      exportVersion: v2.0.2
                      source: World Cube Association Results Export
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '503':
          $ref: '#/components/responses/DatasetUnavailable'
      x-codeSamples:
        - lang: cURL
          source: |
            curl 'http://speedcube.com.br/api/wca-data/v1/rankings?eventId=333&type=single&region=world'
  /wca-data/v1/speedcubers/top:
    get:
      tags: [Rankings]
      summary: List top speedcubers
      operationId: listTopSpeedcubers
      description: Shortcut for world rankings by event and type.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: eventId
          in: query
          required: true
          description: WCA event id.
          schema:
            type: string
            example: '333'
        - name: type
          in: query
          required: true
          description: Ranking type.
          schema:
            type: string
            enum: [average, single]
            example: average
      responses:
        '200':
          description: Top world ranking entries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RankListResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '503':
          $ref: '#/components/responses/DatasetUnavailable'
  /wca-data/v1/results:
    get:
      tags: [Results]
      summary: List results
      operationId: listResults
      description: |
        Lists result rows. Filter by competition, event, and person id to retrieve the subset needed by an application.

        Result values are raw WCA integers. For timed events, values are centiseconds. For Fewest Moves, values are move counts or WCA scaled averages. For multi-blind, values follow WCA multi-blind encoding.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: competitionId
          in: query
          description: Filter by WCA competition id.
          schema:
            type: string
            example: WorldChampionship2023
        - name: eventId
          in: query
          description: Filter by WCA event id.
          schema:
            type: string
            example: '333'
        - name: personId
          in: query
          description: Filter by WCA person id.
          schema:
            type: string
            example: 2012PARK03
      responses:
        '200':
          description: Result entries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultListResponse'
              examples:
                results:
                  value:
                    data:
                      - competitionId: BrizZonSylwesterOpen2022
                        eventId: clock
                        personId: 2013ROGA02
                        position: 1
                        best:
                          raw: 539
                        average:
                          raw: 564
                        solves:
                          - raw: 587
                          - raw: 615
                          - raw: 543
                          - raw: 539
                          - raw: 561
                        round: Final
                        format: Average of 5
                        regionalSingleRecord: null
                        regionalAverageRecord: null
                    pagination:
                      page: 1
                      pageSize: 50
                      total: 1
                      hasNextPage: false
                    meta:
                      datasetId: 2026-06-30-v2.0.2
                      exportDate: '2026-06-30T00:00:16.000Z'
                      exportVersion: v2.0.2
                      source: World Cube Association Results Export
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '503':
          $ref: '#/components/responses/DatasetUnavailable'
      x-codeSamples:
        - lang: cURL
          source: |
            curl 'http://speedcube.com.br/api/wca-data/v1/results?competitionId=WorldChampionship2023&eventId=333'
  /wca-data/v1/scrambles:
    get:
      tags: [Scrambles]
      summary: List scrambles
      operationId: listScrambles
      description: |
        Lists official WCA scramble strings from the active dataset. Filter by competition and event for the most common use case, and add round type, group, or extra-scramble filters when reconstructing a specific scramble set.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: competitionId
          in: query
          description: Filter by WCA competition id.
          schema:
            type: string
            example: WorldChampionship2023
        - name: eventId
          in: query
          description: Filter by WCA event id.
          schema:
            type: string
            example: '333'
        - name: roundTypeId
          in: query
          description: Filter by WCA round type id.
          schema:
            type: string
            example: f
        - name: groupId
          in: query
          description: Filter by scramble group id.
          schema:
            type: string
            example: A
        - name: isExtra
          in: query
          description: Filter regular scrambles (`false`) or extra scrambles (`true`).
          schema:
            type: boolean
            example: false
      responses:
        '200':
          description: Scramble entries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScrambleListResponse'
              examples:
                scrambles:
                  value:
                    data:
                      - id: 1
                        competitionId: FixtureOpen2026
                        eventId: '333'
                        roundTypeId: f
                        groupId: A
                        isExtra: false
                        scrambleNumber: 1
                        scramble: R U R' U'
                    pagination:
                      page: 1
                      pageSize: 50
                      total: 1
                      hasNextPage: false
                    meta:
                      datasetId: 2026-06-30-v2.0.2
                      exportDate: '2026-06-30T00:00:16.000Z'
                      exportVersion: v2.0.2
                      source: World Cube Association Results Export
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '503':
          $ref: '#/components/responses/DatasetUnavailable'
      x-codeSamples:
        - lang: cURL
          source: |
            curl 'http://speedcube.com.br/api/wca-data/v1/scrambles?competitionId=WorldChampionship2023&eventId=333&pageSize=100'
components:
  parameters:
    Page:
      name: page
      in: query
      description: 1-indexed page number.
      schema:
        type: integer
        minimum: 1
        default: 1
        example: 1
    PageSize:
      name: pageSize
      in: query
      description: Number of items per page. The maximum is 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
        example: 50
  responses:
    InvalidRequest:
      description: The request query or path parameters are invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: invalid_request
              message: Invalid request parameters
              requestId: req-123
    NotFound:
      description: The requested item was not found in the active dataset.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: not_found
              message: WCA person not found
              requestId: req-123
    DatasetUnavailable:
      description: The API is healthy, but no active dataset is available.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: dataset_unavailable
              message: WCA Data API dataset is unavailable
              requestId: req-123
  schemas:
...
    EventListResponse:
      allOf:
        - $ref: '#/components/schemas/ListEnvelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Event'
    CountryListResponse:
      allOf:
        - $ref: '#/components/schemas/ListEnvelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Country'
    ContinentListResponse:
      allOf:
        - $ref: '#/components/schemas/ListEnvelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Continent'
    FormatListResponse:
      allOf:
        - $ref: '#/components/schemas/ListEnvelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Format'
    RoundTypeListResponse:
      allOf:
        - $ref: '#/components/schemas/ListEnvelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/RoundType'
    ChampionshipListResponse:
      allOf:
        - $ref: '#/components/schemas/ListEnvelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Championship'
    CompetitionListResponse:
      allOf:
        - $ref: '#/components/schemas/ListEnvelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Competition'
    CompetitionResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: '#/components/schemas/Competition'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    PersonListResponse:
      allOf:
        - $ref: '#/components/schemas/ListEnvelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Person'
    PersonResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: '#/components/schemas/Person'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    RankListResponse:
      allOf:
        - $ref: '#/components/schemas/ListEnvelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/RankEntry'
...